+++ home/tests/test_api.py
@@ -5,6 +5,7 @@ from pathlib import Path
 import pytest
 from bs4 import BeautifulSoup
 from django.contrib.contenttypes.models import ContentType
+from django.core.exceptions import ValidationError  # type: ignore
 from django.core.files.base import File  # type: ignore
 from django.core.files.images import ImageFile  # type: ignore
 from django.urls import reverse
@@ -15,7 +16,12 @@ from wagtail.models import Workflow, WorkflowContentType
 from wagtailmedia.models import Media  # type: ignore
 
 from home.content_import_export import import_content
-from home.models import ContentPage, HomePage, OrderedContentSet, PageView
+from home.models import (
+    ContentPage,
+    HomePage,
+    OrderedContentSet,
+    PageView,
+)
 
 from .page_builder import (
     MBlk,
@@ -1235,3 +1241,19 @@ class TestOrderedContentSetAPI:
         content_str = response.content.decode("utf-8")
         assert "/admin/snippets/home/orderedcontentset/" in content_str
         assert response.status_code == 200
+
+    def test_valid_document_extension(self):
+        """
+        Upload an invalid document type
+        """
+        document = Document.objects.create(
+            title="Example Document", file="invalid_file.exe"
+        )
+
+        with pytest.raises(ValidationError) as validation_error:
+            document.full_clean()
+
+        assert (
+            validation_error.value.messages[0]
+            == "File extension “exe” is not allowed. Allowed extensions are: doc, docx, xls, xlsx, ppt, pptx, pdf, txt."
+        )