+++ CHANGELOG.md
@@ -14,7 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Increased SMS limit to 459 characters
-- Validate character limit fir footer and list items when importing 
+- Validate character limit fir footer and list items when importing
+- Add Document upload extension validation
 
 -->
 ## v1.1.0